fix(verify-standards): break second circular import by moving to api-indexer/
Same pattern as #162 (interpret.ts). After publishing bitbadges@0.34.2 and
bumping the indexer dep, the jest run surfaced another identical
"Super expression must either be null or a function" from a different
entry point:

  src/index.ts -> core/index.ts -> core/verify-standards.ts
    -> api-indexer/docs-types/docs.ts (mid-evaluation; class extends
       something still undefined because of the re-entered core/)

Root cause: verify-standards.ts does `new CollectionDoc(...)` at runtime
— a live dependency from core/ into api-indexer/, same as what interpret.ts
was doing.

Fix: move src/core/verify-standards.ts -> src/api-indexer/verify-standards.ts
(the natural home, since it instantiates an api-indexer class). Update
relative imports, re-exports, and two direct importers in core/.

Scanned the rest of the codebase for the same anti-pattern — this was the
only remaining core/*.ts file with a runtime (non-type) import from
api-indexer/. No other files affected.

Verification:
- `bun run build` passes (tsc + tsc-alias + fixup + dep-check all green)
- `madge --circular` reports: "No circular dependency found!"
- CJS load test: 767 exports, 716 callables probed, 0 broken (previously
  the "Super expression" error broke any import path that touched
  core/index.ts)
- Specific import probe for the failing indexer specs
  (MAINNET_COINS_REGISTRY, TESTNET_COINS_REGISTRY, BitBadgesApiRoutes,
  GetSignInChallengeSuccessResponse, SiwbbChallengeParams, UintRangeArray)
  all resolve cleanly.

Public API is unchanged: top-level `import { verifyStandardsCompliance }
from 'bitbadges'` still works via the api-indexer/index.ts re-export.
