+++ packages/bitbadgesjs-sdk/src/api-indexer/index.ts
@@ -8,5 +8,6 @@ export * from './BitBadgesCollection.js';
 export * from './BitBadgesUserInfo.js';
 // Moved here from core/ to break a runtime circular import (see comment in core/index.ts).
 export * from './interpret.js';
+export * from './verify-standards.js';
 export * from './requests/index.js';
 export * from './responses/index.js';
