+++ packages/bitbadgesjs-sdk/src/core/index.ts
@@ -62,7 +62,9 @@ export {
 export * from './validate.js';
 export * from './blankCriteria.js';
 export * from './audit.js';
-export * from './verify-standards.js';
+// NOTE: verify-standards.js moved to ../api-indexer/verify-standards.js — same reason as interpret.js.
+// It creates a runtime `new CollectionDoc(...)`, which requires docs-types/docs.js — and that chain
+// used to reach back into core/ mid-load, producing a "Super expression" error.
 export * from './review.js';
 
 export * from './builders/index.js';
